pyproject.toml
[project]
name = "edge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "insightface>=0.7.3",
    "numpy>=2.4.4",
    "onnxruntime>=1.24.4",
    "opencv-python>=4.13.0.92",
    "psycopg2[binary]>=2.9.12",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "qrcode>=8.2",
    "requests>=2.33.1",
]
